wifi: Update bcm driver to 101.10.591.52.27 to fix CTS [2/2]
PD#SWPL-154900

Problem:
Because TRUNK was debugged in an Android 64bit + kernel 64-bit
environment and was not debugged in an Android 32-bit +
kernel 64bit environment,
this issue occurred due to compatibility problems.

Solution:
It is necessary to update the WiFi driver to fix the compatibility
problem.
WiFi Driver information 101.10.591.52.27 (20240306-1)

Verify:
tm2+t7c

Change-Id: Ibbea9b37084a1eafdec6ada7fce7e65e548795e2
Signed-off-by: li.dong <li.dong@amlogic.com>